chore: refactor budget feature

+++ src/hooks/queries/useBudgets.ts
@@ -1,7 +1,8 @@
-
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { supabase } from "@/integrations/supabase/client";
 import { useAuth } from "@/hooks/useAuth";
+import { useToast } from "@/hooks/use-toast";
+import { BudgetFormData } from "@/form-dto/budget";
 
 export const useBudgets = () => {
   const { user } = useAuth();
@@ -24,18 +25,97 @@ export const useBudgets = () => {
 
 export const useDeleteBudget = () => {
   const queryClient = useQueryClient();
+  const { toast } = useToast();
+  const { user } = useAuth();
 
   return useMutation({
     mutationFn: async (id: number) => {
       const { error } = await supabase
         .from("budgets")
         .delete()
+        .eq("user_id", user?.id)
         .eq("id", id);
 
       if (error) throw error;
     },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ["budgets"] });
+      toast({
+        title: "Berhasil",
+        description: "Budget berhasil dihapus",
+      });
+    },
+    onError: (error: any) => {
+      toast({
+        title: "Error",
+        description: error.message,
+        variant: "destructive",
+      });
+    },
+  });
+};
+
+export const useCreateBudget = () => {
+  const queryClient = useQueryClient();
+  const { user } = useAuth();
+  const { toast } = useToast();
+
+  return useMutation({
+    mutationFn: async (newBudget: BudgetFormData) => {
+      const { error } = await supabase
+        .from("budgets")
+        .insert({
+          ...newBudget,
+          user_id: user.id,
+        });
+
+      if (error) throw error;
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ["budgets"] });
+      toast({
+        title: "Berhasil",
+        description: "Budget berhasil ditambahkan",
+      });
+    },
+    onError: (error) => {
+      toast({
+        title: "Error",
+        description: `Gagal menambahkan budget: ${error.message}`,
+        variant: "destructive",
+      });
+    },
+  });
+};
+
+export const useUpdateBudget = () => {
+  const queryClient = useQueryClient();
+  const { toast } = useToast();
+  const { user } = useAuth();
+
+  return useMutation({
+    mutationFn: async ({ id, ...budget }: BudgetFormData & { id: number }) => {
+      const { error } = await supabase
+        .from("budgets")
+        .update(budget)
+        .eq("user_id", user?.id)
+        .eq("id", id);
+
+      if (error) throw error;
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ["budgets"] });
+      toast({
+        title: "Berhasil",
+        description: "Budget berhasil diperbarui",
+      });
+    },
+    onError: (error) => {
+      toast({
+        title: "Error",
+        description: `Gagal memperbarui budget: ${error.message}`,
+        variant: "destructive",
+      });
     },
   });
 };
